+++ web/packages/plugins/analysis-platform/src/modules/Analysis/components/AnalysisTable.tsx
@@ -1,6 +1,8 @@
-import React from 'react'
-import { DynamicTable, formatDate } from '@dmt/common'
-import { TAnalysis } from '../Types'
+import React, { useContext } from 'react'
+import { ApplicationContext, DynamicTable, formatDate } from '@dmt/common'
+import { TAnalysis } from '../../../Types'
+import { useParams } from 'react-router-dom'
+import { DEFAULT_DATASOURCE_ID } from '../../../const'
 
 const columns: Array<string> = [
   'Analysis name',
@@ -17,6 +19,8 @@ type TAnalysisRow = {
   created: Date | string
   updated: Date | string
   creator: string
+  index: number
+  url: string
 }
 
 type TAnalysisTableProps = {
@@ -24,16 +28,31 @@ type TAnalysisTableProps = {
 }
 
 const onRowClicked = (event: any) => {
-  const index = event.target.parentElement.tabIndex
   //@ts-ignore
-  document.location = `${document.location.pathname}.analyses.${index}`
+  document.location = event.target.parentElement.accessKey
 }
 
 export const AnalysisTable = (props: TAnalysisTableProps) => {
   const { analyses } = props
+  const { data_source, entity_id } = useParams<{
+    data_source: string
+    entity_id: string
+  }>()
+  const settings = useContext(ApplicationContext)
+
+  const getViewUrl = (index: number, rowDocumentId: string): string => {
+    const urlBase = `/${settings.urlPath}/view`
+
+    if (data_source && entity_id) {
+      // We're viewing an asset. Access the .analyses subdocuments
+      return `${urlBase}/${data_source}/${entity_id}.analyses.${index}`
+    }
+
+    return `${urlBase}/${DEFAULT_DATASOURCE_ID}/${rowDocumentId}`
+  }
 
   const rows: Array<TAnalysisRow> = []
-  analyses?.forEach((analysis) => {
+  analyses?.forEach((analysis, index) => {
     const row: TAnalysisRow = {
       _id: analysis._id,
       name: analysis.label || analysis.name,
@@ -41,6 +60,8 @@ export const AnalysisTable = (props: TAnalysisTableProps) => {
       created: formatDate(analysis.created),
       updated: formatDate(analysis.updated),
       creator: analysis.creator,
+      index: index,
+      url: getViewUrl(index, analysis._id),
     }
     rows.push(row)
   })
