+++ web/packages/plugins/analysis-platform/src/modules/Analysis/components/CreateAnalysisButton.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Link, useLocation, useParams } from 'react-router-dom'
 import { Button } from '@equinor/eds-core-react'
 
-import { CreateAnalysisButtonProps } from '../Types'
+import { CreateAnalysisButtonProps } from '../../../Types'
 
 export const CreateAnalysisButton = (props: CreateAnalysisButtonProps) => {
   const { urlPath } = props
