+++ web/packages/plugins/analysis-platform/src/modules/Asset/index.tsx
@@ -1,3 +1,4 @@
 export * from './components'
 export * from './AssetOverview'
 export * from './AssetCreate'
+export * from './AssetView'
