--- web/packages/plugins/analysis/jest.config.js
@@ -1,8 +1,0 @@
-const base = require('../../../jest.config.base')
-const packageJson = require('./package.json')
-
-module.exports = {
-  ...base,
-  name: packageJson.name,
-  displayName: packageJson.name,
-}
