--- web/packages/plugins/analysis/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@dmt/analysis",
-  "version": "1.0.0",
-  "main": "src/index.tsx",
-  "dependencies": {
-    "styled-components": "^5.3.5"
-  }
-}
