--- web/packages/plugins/analysis/src/components/Design/Icons.ts
@@ -1,18 +1,0 @@
-import { Icon } from '@equinor/eds-core-react'
-import {
-  add,
-  assignment_user,
-  close,
-  edit_text,
-  time,
-} from '@equinor/eds-icons'
-
-Icon.add({
-  add,
-  assignment_user,
-  close,
-  edit_text,
-  time,
-})
-
-export default Icon
