--- web/packages/plugins/analysis/src/components/Design/Styled.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-
-export const OnRight = styled.div`
-  display: flex;
-  margin-right: 50px;
-`
