--- web/packages/plugins/analysis/src/components/index.tsx
@@ -1,4 +1,0 @@
-export * from './AnalysisInfo'
-export * from './AnalysisTable'
-export * from './AnalysisJobTable'
-export * from './AssetInfo'
