--- web/packages/plugins/analysis/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "../../../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "rootDir": "src"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "src/**/*.spec.ts", "dist"]
-}
